chore(package-lock): update dependencies and add new packages
This commit updates several package versions in package-lock.json, including @jridgewell/sourcemap-codec, @types/node, dotenv, and pg, among others. Additionally, new packages such as available-typed-arrays, call-bind, define-data-property, for-each, has-property-descriptors, is-callable, is-typed-array, which-typed-array, and others are added to enhance functionality and maintain compatibility with the latest versions. These changes ensure the project remains up-to-date with dependency management.